Revert "Custom ocr fix replace list"
