+++ src/SubtitleEdit.csproj
@@ -783,7 +783,6 @@
     <Compile Include="Logic\ColorChooser\ColorChangedEventArgs.cs" />
     <Compile Include="Logic\ColorChooser\ColorHandler.cs" />
     <Compile Include="Logic\ColorChooser\ColorWheel.cs" />
-    <Compile Include="Logic\OcrFixReplaceList.cs" />
     <Compile Include="Logic\Extensions.cs" />
     <Compile Include="Logic\FileUtils.cs" />
     <Compile Include="Logic\Forms\CheckForUpdatesHelper.cs" />
