+++ src/Test/Test.csproj
@@ -46,7 +46,6 @@
     <Compile Include="FixCommonErrorsTest.cs" />
     <Compile Include="languageTest.cs" />
     <Compile Include="Logic\HtmlUtilsTest.cs" />
-    <Compile Include="Logic\OcrFixReplaceListTest.cs" />
     <Compile Include="RemoveTextForHearImpairedTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SubtitleFormatsTest.cs" />
